# contact/contact.go
package contact

type Contact struct {
	ID        string `json:"id"`
	FirstName string `json:"firstName"`
	LastName  string `json:"lastName"`
	Phone     string `json:"phone"`
	Address   string `json:"address"`
}

type Filters struct {
	FullText string
	Limit    int
	Offset   int
}